Language.Haskell.Exts.ExactPrint

Description

Exact-printer for Haskell abstract syntax. The input is a (semi-concrete) abstract syntax tree, annotated with exact source information to enable printing the tree exactly as it was parsed.

Documentation

exactPrint :: ExactP ast => ast SrcSpanInfo -> [Comment] -> String

Print an AST exactly as specified by the annotations on the nodes in the tree.

class Annotated ast => ExactP ast

Minimal complete definition

exactP