Language.Haskell.Exts.Pretty

Description

Pretty printer for Haskell with extensions.

Pretty printing

class Pretty a

Things that can be pretty-printed, including all the syntactic objects in Language.Haskell.Exts.Syntax.

prettyPrintStyleMode :: Pretty a => Style -> PPHsMode -> a -> String

render the document with a given mode. renderWithMode :: PPHsMode -> Doc -> String renderWithMode = renderStyleMode P.style

render the document with defaultMode. render :: Doc -> String render = renderWithMode defaultMode

pretty-print with a given style and mode.

prettyPrintWithMode :: Pretty a => PPHsMode -> a -> String

pretty-print with the default style and a given mode.

prettyPrint :: Pretty a => a -> String

pretty-print with the default style and defaultMode.

Pretty-printing styles (from Text.PrettyPrint.HughesPJ)

data Style

A rendering style. Allows us to specify constraints to choose among the many different rendering options.

Constructors

Style

Fields

mode :: Mode

The rendering mode.

lineLength :: Int

Maximum length of a line, in characters.

ribbonsPerLine :: Float

Ratio of line length to ribbon length. A ribbon refers to the characters on a line excluding indentation. So a lineLength of 100, with a ribbonsPerLine of 2.0 would only allow up to 50 characters of ribbon to be displayed on a line, while allowing it to be indented up to 50 characters.

style :: Style

The default style (mode=PageMode, lineLength=100, ribbonsPerLine=1.5).

data Mode

Rendering mode.

Constructors

PageMode
Normal rendering (lineLength and ribbonsPerLine respected').

ZigZagMode
With zig-zag cuts.

LeftMode
No indentation, infinitely long lines (lineLength ignored), but explicit new lines, i.e., text "one" $$ text "two", are respected.

OneLineMode
All on one line, lineLength ignored and explicit new lines ($$) are turned into spaces.

Haskell formatting modes

data PPHsMode

Pretty-printing parameters.

Note: the onsideIndent must be positive and less than all other indents.

Constructors

PPHsMode

Fields

classIndent :: Indent

indentation of a class or instance

doIndent :: Indent

indentation of a do-expression

multiIfIndent :: Indent

indentation of the body of a case expression

caseIndent :: Indent

indentation of the body of a multi-if expression

letIndent :: Indent

indentation of the declarations in a let expression

whereIndent :: Indent

indentation of the declarations in a where clause

onsideIndent :: Indent

indentation added for continuation lines that would otherwise be offside

spacing :: Bool

blank lines between statements?

layout :: PPLayout

Pretty-printing style to use

linePragmas :: Bool

add GHC-style LINE pragmas to output?

type Indent = Int

data PPLayout

Varieties of layout we can use.

Constructors

PPOffsideRule
classical layout

PPSemiColon
classical layout made explicit

PPInLine
inline decls, with newlines between them

PPNoLayout
everything on a single line

defaultMode :: PPHsMode

The default mode: pretty-print using the offside rule and sensible defaults.

Primitive Printers

prettyPrim :: Pretty a => a -> Doc

pretty-print with the default style and defaultMode.

prettyPrimWithMode :: Pretty a => PPHsMode -> a -> Doc

pretty-print with the default style and a given mode.